Changed Rating Value from Float to Int to be more relevant (on scale of 1-5) and applied the corresponding modifications to the Mapper Class

+++ KitLapBackend/DTOs/Requests/AddRatingDto.cs
@@ -11,6 +11,6 @@ namespace KitLapBackend.DTOs.Requests
         [Required]
         public int ProductId { get; set; }
         [Required, Range(0,5)]
-        public float RatingValue { get; set; }
+        public int RatingValue { get; set; }
     }
 }
